refacotr: resize strategy 패키지 분리
