+++ imagestorage/src/main/java/com/woowacourse/imagestorage/strategy/resize/ImageResizeStrategy.java
@@ -1,4 +1,4 @@
-package com.woowacourse.imagestorage.strategy;
+package com.woowacourse.imagestorage.strategy.resize;
 
 public interface ImageResizeStrategy {
